Add relational and logical operators + refactor binop parser
We now support OR and AND with short circuiting! (Yet to be tested
since we don't yet have local variables to play with).

The binop parser took a bit of an overhaul factoring out the common
code so that it's easier to describe the operator precendence
relationships without being overly repetitive.