Add support for pointers! (tests missing)
This commit adds initial support for taking pointers / dereferencing.
The type system is still a bit of a hot mess, so all type information
is actually not looked at, but the functionality still seems to be
there.

Still need to add some tests for pointers/dereferencing to ensure that
it works in some edge cases as well.