Add helper to create builtins for syscalls + implement `read()`
This was possible, but very tedious to do by hand before. Now we
automate it based on the number of arguments.

Note that currently we can't just add `syscall3()` etc as builtins
because the actual numbers for the system calls vary from one system
to another, and we want to maintain support for macOS and Linux (at
least for now).