Add support for basic structs
Structs for now (and probably for the near future) are not allowed
to be passed by value, and instead you just pass a pointer to it.

Nested structs can also be defined, and they can be either anonymous,
or named (in which case only the members can access the type).