[C]: Add `here` keyword that evaluates to a string with it's location
It's not really a keyword like the other ones, but just handled
completely at the lexer level since it already knows the location of the
token, so it injects a string literal instead.

We also use this in the self-hosted compiler now for better error
reporting for where the error happened internally.