Escape strings / chars manually and output ascii values in nasm
Previously we were using the escape-string ability in NASM that
allowed strings to be enclosed in backticks. However, this feature
is not really supported by other assemblers, so in the interest
of portability we now escape the char/string literals manually.

This obviously increases the size of the generated assembly file,
but it allows us to move away from NASM (which is getting very slow)