Fix error return values for syscalls on macOS
Unlike Linux, it seems that on macOS that syscalls do not return
negative values on failure, but rather return the positive error
code and set the Carry Flag when there is an error.

This commit adds some extra code to negate the error code if the
carry flag is set on macOS. With this, at least the `open()` syscall
seems to actually return an error when a non-existent file is opened.